massive refactor for nav bar
